+++ dataprofiler/tests/profilers/profiler_options/test_profiler_presets.py
@@ -1,9 +1,11 @@
 import unittest
+from collections import defaultdict
 from unittest import mock
 
 from dataprofiler import ProfilerOptions
 from dataprofiler.labelers.base_data_labeler import BaseDataLabeler
 from dataprofiler.plugins.__init__ import getPlugins
+from dataprofiler.plugins.decorators import plugin_decorator, plugins_dict
 
 
 @mock.patch(
@@ -36,7 +38,16 @@ class TestProfilerPresets(unittest.TestCase):
         self.assertTrue(options.structured_options.order.is_enabled)
 
     def test_plugin_presets(self, *mocks):
-        test_dict = getPlugins("test_preset")
-        if test_dict is not None:
-            for plugin_preset in test_dict:
-                test_dict[plugin_preset](self, *mocks)
+        mock_plugin_execution = mock.Mock()
+        with mock.patch.dict(plugins_dict, defaultdict(dict)) as mock_plugin_dict:
+
+            @plugin_decorator(typ="test_preset", name="mock_test")
+            def test_plugin():
+                mock_plugin_execution()
+
+            expected_default_dict = defaultdict(dict)
+            expected_default_dict["test_preset"]["mock_test"] = test_plugin
+            self.assertDictEqual(expected_default_dict, mock_plugin_dict)
+
+            test_get_dict = getPlugins("test_preset")
+            self.assertDictEqual({"mock_test": test_plugin}, test_get_dict)
